+++ sketch/src/androidTest/java/com/github/panpf/sketch/test/decode/GifAnimatedDrawableDecoderTest.kt
@@ -176,7 +176,7 @@ class GifAnimatedDrawableDecoderTest {
                 Assert.assertEquals(LOCAL, this.dataFrom)
                 Assert.assertNull(this.transformedList)
                 val animatedImageDrawable =
-                    ((this.drawable as SketchAnimatableDrawable).wrappedDrawable as ScaledAnimatedImageDrawable).drawable
+                    ((this.drawable as SketchAnimatableDrawable).drawable as ScaledAnimatedImageDrawable).drawable
                 Assert.assertEquals(-1, animatedImageDrawable.repeatCount)
             }
 
@@ -192,7 +192,7 @@ class GifAnimatedDrawableDecoderTest {
                 Assert.assertEquals(LOCAL, this.dataFrom)
                 Assert.assertEquals(listOf(createInSampledTransformed(2)), this.transformedList)
                 val animatedImageDrawable =
-                    ((this.drawable as SketchAnimatableDrawable).wrappedDrawable as ScaledAnimatedImageDrawable).drawable
+                    ((this.drawable as SketchAnimatableDrawable).drawable as ScaledAnimatedImageDrawable).drawable
                 Assert.assertEquals(3, animatedImageDrawable.repeatCount)
             }
     }